+++ src/CommerceCoreServiceProvider.php
@@ -5,6 +5,7 @@ namespace Lalalili\CommerceCore;
 use Lalalili\CommerceCore\Services\EntitlementService;
 use Lalalili\CommerceCore\Services\OrderLifecycleHookDispatcher;
 use Lalalili\CommerceCore\Services\OrderLifecycleService;
+use Lalalili\CommerceCore\Services\PaymentLogService;
 use Lalalili\CommerceCore\Support\ModelAttributeMapper;
 use Lalalili\CommerceCore\Support\OrderItemNormalizer;
 use Lalalili\CommerceCore\Support\OrderNumberGenerator;
@@ -31,5 +32,6 @@ class CommerceCoreServiceProvider extends PackageServiceProvider
         $this->app->singleton(EntitlementService::class);
         $this->app->singleton(OrderLifecycleHookDispatcher::class);
         $this->app->singleton(OrderLifecycleService::class);
+        $this->app->singleton(PaymentLogService::class);
     }
 }